Refactor, update, simplify package
Update to src/ layout with pyproject.toml. Remove cruft configuration files. Update package dependencies based on what's actually imported.
