+++ .github/workflows/pythonpackage.yaml
@@ -28,10 +28,10 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           cache: 'pip'
-          cache-dependency-path: 'requirements_dev.txt'
+          cache-dependency-path: 'requirements.txt'
       - name: Install dependencies
         run: |
-          pip install -r requirements_dev.txt
+          pip install -r requirements.txt
       - name: Format check with flake8
         run: |
           flake8
@@ -51,7 +51,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           cache: 'pip'
-          cache-dependency-path: 'requirements_dev.txt'
+          cache-dependency-path: 'requirements.txt'
       - name: Install dependencies
         run: |
           pip install -r requirements_dev.txt
@@ -60,7 +60,7 @@ jobs:
           pip install . --no-deps
       - name: Test with pytest
         run: |
-          pytest -v --cov impactlab_tools --cov-report term-missing --cov-report xml
+          pytest -v --cov-report term-missing --cov-report xml
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3
