--- MANIFEST.in
@@ -1,10 +1,0 @@
-
-include README.rst
-
-include requirements*
-
-recursive-include tests *
-recursive-exclude * __pycache__
-recursive-exclude * *.py[co]
-recursive-include impactlab_tools assets *.csv *.nc *.dbf *.prj *.shp *.shx
-recursive-include docs *.rst conf.py Makefile make.bat *.jpg *.png *.gif
