--- pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-testpaths = impactlab_tools tests docs
-addopts = --cov=impactlab_tools --cov=tests --doctest-modules --cov-report term-missing
