--- requirements_dev.txt
@@ -1,20 +1,0 @@
-numpy
-netcdf4
-xarray
-pandas
-scipy
-bottleneck
-geopandas
-dask
-matplotlib
-Sphinx
-sphinx-rtd-theme
-pip
-wheel
-flake8
-pytest
-coverage
-pytest-cov
-pytest-runner
-coveralls
-pyyaml
