--- requirements_rtd.txt
@@ -1,3 +1,0 @@
--r requirements.txt
--r requirements_conda.txt
--r requirements_test.txt
