--- requirements_test.txt
@@ -1,10 +1,0 @@
-Sphinx
-sphinx-rtd-theme
-pip
-wheel
-flake8
-pytest
-coverage
-pytest-cov
-pytest-runner
-coveralls
