+++ setup.cfg
@@ -1,26 +1,6 @@
-[bumpversion]
-current_version = 0.4.0
-commit = True
-tag = True
-
-[bumpversion:file:setup.py]
-search = version='{current_version}'
-replace = version='{new_version}'
-
-[bumpversion:file:impactlab_tools/__init__.py]
-search = __version__ = '{current_version}'
-replace = __version__ = '{new_version}'
-
-[bdist_wheel]
-universal = 1
-
 [flake8]
 exclude = docs,build,dist
 ignore = E203,E266,E402,E501,W503,F401,C901
 max-line-length = 88
 max-complexity = 18
 select = B,C,E,F,W,T4,B9
-
-[aliases]
-test = pytest
-
