+++ src/impactlab_tools/__init__.py
@@ -5,7 +5,6 @@
 from __future__ import absolute_import
 
 __author__ = """Climate Impact Lab"""
-__version__ = '0.4.0'
 
 _module_imports = (
 )
