+++ src/impactlab_tools/acp/aggregate.py
@@ -4,8 +4,6 @@ from __future__ import absolute_import
 import xarray as xr
 import os
 
-import impactlab_tools.assets
-
 
 def population_weighted_mean(
         ds,
