+++ src/impactlab_tools/assets/GCP_impact_regions.nc
