+++ src/impactlab_tools/utils/binning.py
