+++ tests/utils/test_configdict.py
@@ -1,6 +1,6 @@
 import pytest
 
-from impactlab_tools.utils.configdict import ConfigDict, gather_configtree
+from src.impactlab_tools.utils.configdict import ConfigDict, gather_configtree
 
 
 @pytest.fixture
